+++ CHANGELOG.md
@@ -3,4 +3,19 @@ All notable changes to this project will be documented in this file.
 
 ## [0.0.1 - 2022-08-31]
 ### Added
-- **Implemented all Administration Endpoints (`/rest/api/latest/admin/`)**:
+- **Implemented all Permission Endpoints (`/rest/api/latest/`)**:
+
+    - `/rest/api/latest/mypermissions`
+    - `/rest/api/latest/permissions`
+
+---
+
+- **Implemented all Application Properties Endpoints (`/rest/api/latest/application-properties`)**:
+
+    - `/rest/api/latest/application-properties`
+    - `/rest/api/latest/application-properties/{id}`
+    - `/rest/api/latest/application-properties/advanced-settings`
+
+---
+
+